requirements.txt
pandas==2.2.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
selenium==4.18.1
webdriver-manager==4.0.1